Hitting weather API endpoint
Summary: Added Dagger to CityNameController. Also hit the weather api endpoint by using the get method and query. Retrofit was used to get the connection to the endpoint.

Test Plan:
Patch firstProject
launch project
Click on main button and verify that a toast will appear in the second activity(bottom of the screen)

Reviewers: sinistro831

Reviewed By: sinistro831

Differential Revision: https://kr-is-team.phacility.com/D13